Setup the JS file with the Book and Library constructors and methods
A 'Book' constructor has been setup to represent a single book.
Methods to print the book info have also been coded.
The date in which the Book is created is also included, as well as
its state code, represented by an integer.
Note that the possible states are defined in the global array:

	let statesOfRead = ['not read yet', 'just started', 'about half read',
'almost finished','read'];

and the current state is indexed by the state code (0,1,2,3,4).

A 'Library' constructor has been setup to represent a library, whose
books are described in the this.books array of Book objects.
Methods to print the library info (books) have also been coded.

A function to add a book object to the library has also been included.
The function should check if the book is already present in the library,
and in that case ignore the request. Note that currently this is simply
done by:

	this.books.includes(book)

hence books which have the same information but different
times of inclusion can both be present (they are different objects).

Finally, 40 sample books information have been added to populate the
array of books. A function that initialize the library to a given number
of books has been coded.