Add function to generate a generic book box programmatically
The function requires a Book object, and its data are used to
populate the book box elements.

Note: the functionalities regarding the pages read so far and the
current state are not implemented yet. Some dummy data have been used.

Also, the book box dimensions are not fixed yet (their size now
depend on the number of rows which the title spans).