Adapt font size of the books' titles to fit their containers
This is done both at after the initialization, and when the window
is resized (because the size of the box describing the books can
change).

The height of the div containing the title of each book is set to
a fixed value, fixing the (maximum) number of rows we allow.
Then, we chack overflow of such text and if that occurs, we reduce
the font size.