Add the possibility to increase/decrease the read pages faster
Previously, a click/tap was required to increase/decrease the
number of read pages. Keeping the corresponding button pressed did
nothing.

Now, when you keep such buttons pressed, the number of read pages
continues to increase/decrease until the button is unpressed.

An update to the number of read pages is done approximately every
200ms. This is done explotitng the setInterval/clearInterval
functions, so the timing between such events might not be exact.