Add some icons to the footer of the book boxes
A 'Like' icon is used to mark a book as favourite. Its state toggles
between two values, and the icon changes if the book is a favourite
one.

A 'Share' icon is just a link to a placeholder link.

A 'Delete' icon will be used to delete the corresponding book. This
functionality is not implemented yet.