Add the functionality to add a new book (partial)
Currently, data are not read from the form, yet: a sample book is added, instead.
Specific functions have been written to handle the addition of a new book.

A small fix has been also applied to the event listeners for the increase / decrease read pages functionality.
A 'pointerout' event was , causing the associated callback (which, e.g., checks if the books ordering has changed, and so on) to be unnecessarily executed when the mouse just hovers over such buttons.

First, the 'pointerout' event listeners have been substituted with the 'pointerleave' event listeners.
Then, a custom event listerer has been added for such events, which calls the actual handler only if the pointer is not a mouse or, if it is a mouse, a button is being pressed.