Add functionality to add new books using form data filled by user
Note that the form validation is not completed yet.
Form data should also be reset after submitting, which is missing
as of now.

Also, fix the sorting order for numeric data: push NaN at the
beginning. The year field could be NaN, if the data is not provided.

Finally, fix the visualization for the empty 'genre' and 'year'.
The empty genre is represented by a ''. The empty year by the NaN
(since years are numeric). In both cases, a '-' is shown.