Fix the callback to add a new book
Previously an event listener was used for the 'click' event on the
submit button to add a new book. While it worked, this prevented
the HTML validation to take place.

Now this has been substituted by an event listener for the 'submit'
event on the form itself. Now, when clicking the submit button,
the HTML validation takes place before executing the event handler.

Note that by using 'method="dialog"' on the form element, it is not
necessary to:

	1) prevent default form behaviour, since data are not submitted
	with this method

	2) close the dialog modal, since this is automatically done with
	this method

Still, it is necessary to reset the form fields when closing the
modal. This is done with the .reset() method of the form.

A small fix was also made to the Library.prototype.search method.
Before this, when the string to search was empty "", all items were
matched by default, but the DOM was not properly updated.