Add event listeners for the state and pages inputs in new book form
If the user explicitly clicks on the 'not read' state, the user can
only modify the number of pages. The number of read pages is fixed
to 0. This is also the default initial condition.

Similarly, if the user explicitly clicks on the 'read' state, the
user can only modify the number of pages. Still, the number of read
pages is set to be equal to the number of pages, and changes
accordingly if this gets modified.

If the user clicks on 'reading', it can edit both the number of
pages and the number of read pages. The 'reading' state gets
unchecked if the number of read pages is out of range (ie, larger
than the number of pages). Also it changes to 'not read' if the
number of read pages is 0, or to 'read' if it is equal to the number
of pages.

In all the cases, a input validation is being made for the pages
and read pages inputs. For instance, values are ignored if they are
negative or invalid.