Make the display settings div sticky
The display settings div is now sticky. At the top of the page it
is shown as before.
When the user scrolls down, this disappears. However, when scrolling
up it reappears.
This avoids the user to go back to the top to change such settings.

A throttle function from has been used to limit the burden due to
the continuous events to to handled during scrolling.
Such function has been used also for the resize function.

Also, now the initial default ordering shows the most recent added
books to the top (descend order sort).