Implement the single-word search in a string
Before this modification, the exact string needed to match to show
a book. Now, the the search is done for each single word in the
string, and all of such words must match at least once.

Eg, to search the string 'The of', before this modification the
book 'The Lord of the Rings: The Fellowship of the Ring' would not
match. With this modification, it matches.