Reset the search/favourites/state filters when adding a new book
This allows to always show the newly added book.
The same, is done also when the entire library is cleared.

Finally, the initial book list is randomized, just to show a
different template each time the page is reloaded.
In practice, this should use the user data from local/cloud
storage, which will be implemented in future.